Runtime: extract storage maintenance library
